Title: 자주 등장한 top k 숫자, Time: 76ms, Memory: 3MB, Status: Passed - CodeTree

+++ 240705/자주 등장한 top k 숫자/top-k-frequent-elements.cpp
@@ -24,7 +24,7 @@ int main() {
 
     sort(vect.begin(), vect.end());
 
-    for (int i = vect.size() - 1; i >= vect.size() - k; i--){
+    for (int i = (int)vect.size() - 1; i >= (int)vect.size() - k; i--){
         cout << vect[i].second << " ";
     }
